Add "Update" button and enhance DataGridView properties
Introduced a new "Update" button column to the `dgvCategory`
DataGridView in the `frmCatgory_Load` method. Enhanced the
appearance and functionality of the DataGridView by disabling
user actions like adding, deleting, resizing, and reordering
rows and columns. Customized the appearance with specific
colors, set the background color to white, and configured
column headers and cell selection. Set the DataGridView to
read-only, hid row headers, and enabled auto-sizing for columns.